requirements.txt
openpyxl==3.1.5
openai==2.29.0
cryptography==44.0.1
fastapi==0.117.1
Jinja2==3.1.6
python-multipart==0.0.20
pywebview==5.4
uvicorn==0.36.0
